requirements.txt
Flask
python-docx
openai